Module aws_sdk_snowball::error

Modules

cluster_limit_exceeded_exception

See ClusterLimitExceededException

conflict_exception

See ConflictException

ec2_request_failed_exception

See Ec2RequestFailedException

invalid_address_exception

See InvalidAddressException

invalid_input_combination_exception

See InvalidInputCombinationException

invalid_job_state_exception

See InvalidJobStateException

invalid_next_token_exception

See InvalidNextTokenException

invalid_resource_exception

See InvalidResourceException

kms_request_failed_exception

See KmsRequestFailedException

return_shipping_label_already_exists_exception

See ReturnShippingLabelAlreadyExistsException

unsupported_address_exception

See UnsupportedAddressException

Structs

CancelClusterError

CancelJobError

ClusterLimitExceededException

Job creation failed. Currently, clusters support five nodes. If you have fewer than

ConflictException

You get this exception when you call CreateReturnShippingLabel more than once when other requests are not completed.

CreateAddressError

CreateClusterError

CreateJobError

CreateLongTermPricingError

CreateReturnShippingLabelError

DescribeAddressError

DescribeAddressesError

DescribeClusterError

DescribeJobError

DescribeReturnShippingLabelError

Ec2RequestFailedException

Your IAM user lacks the necessary Amazon EC2 permissions to perform the attempted

GetJobManifestError

GetJobUnlockCodeError

GetSnowballUsageError

GetSoftwareUpdatesError

InvalidAddressException

The address provided was invalid. Check the address with your region's carrier, and try

InvalidInputCombinationException

Job or cluster creation failed. One or more inputs were invalid. Confirm that the

InvalidJobStateException

The action can't be performed because the job's current state doesn't allow that action

InvalidNextTokenException

The NextToken string was altered unexpectedly, and the operation has

InvalidResourceException

The specified resource can't be found. Check the information you provided in your last

KmsRequestFailedException

The provided AWS Key Management Service key lacks the permissions to perform the

ListClusterJobsError

ListClustersError

ListCompatibleImagesError

ListJobsError

ListLongTermPricingError

ReturnShippingLabelAlreadyExistsException

You get this exception if you call CreateReturnShippingLabel and a valid return

UnsupportedAddressException

The address is either outside the serviceable area for your region, or an error

UpdateClusterError

UpdateJobError

UpdateJobShipmentStateError

UpdateLongTermPricingError

Enums

CancelClusterErrorKind

CancelJobErrorKind

CreateAddressErrorKind

CreateClusterErrorKind

CreateJobErrorKind

CreateLongTermPricingErrorKind

CreateReturnShippingLabelErrorKind

DescribeAddressErrorKind

DescribeAddressesErrorKind

DescribeClusterErrorKind

DescribeJobErrorKind

DescribeReturnShippingLabelErrorKind

GetJobManifestErrorKind

GetJobUnlockCodeErrorKind

GetSnowballUsageErrorKind

GetSoftwareUpdatesErrorKind

ListClusterJobsErrorKind

ListClustersErrorKind

ListCompatibleImagesErrorKind

ListJobsErrorKind

ListLongTermPricingErrorKind

UpdateClusterErrorKind

UpdateJobErrorKind

UpdateJobShipmentStateErrorKind

UpdateLongTermPricingErrorKind