Struct aws_sdk_snowball::input::create_job_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateJobInput

Implementations

impl Builder

pub fn job_type(self, input: JobType) -> Self

Defines the type of job that you're creating.

pub fn set_job_type(self, input: Option<JobType>) -> Self

pub fn resources(self, input: JobResource) -> Self

Defines the Amazon S3 buckets associated with this job.

With IMPORT jobs, you specify the bucket or buckets that your transferred data will be imported into.

With EXPORT jobs, you specify the bucket or buckets that your transferred data will be exported from. Optionally, you can also specify a KeyRange value. If you choose to export a range, you define the length of the range by providing either an inclusive BeginMarker value, an inclusive EndMarker value, or both. Ranges are UTF-8 binary sorted.

pub fn set_resources(self, input: Option<JobResource>) -> Self

pub fn on_device_service_configuration(
self,
input: OnDeviceServiceConfiguration
) -> Self

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

pub fn set_on_device_service_configuration(
self,
input: Option<OnDeviceServiceConfiguration>
) -> Self

pub fn description(self, input: impl Into<String>) -> Self

Defines an optional description of this specific job, for example Important Photos 2016-08-11.

pub fn set_description(self, input: Option<String>) -> Self

pub fn address_id(self, input: impl Into<String>) -> Self

The ID for the address that you want the Snow device shipped to.

pub fn set_address_id(self, input: Option<String>) -> Self

pub fn kms_key_arn(self, input: impl Into<String>) -> Self

The KmsKeyARN that you want to associate with this job. KmsKeyARNs are created using the CreateKey AWS Key Management Service (KMS) API action.

pub fn set_kms_key_arn(self, input: Option<String>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The RoleARN that you want to associate with this job. RoleArns are created using the CreateRole AWS Identity and Access Management (IAM) API action.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn snowball_capacity_preference(self, input: SnowballCapacity) -> Self

If your job is being created in one of the US regions, you have the option of specifying what size Snow device you'd like for this job. In all other regions, Snowballs come with 80 TB in storage capacity.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

pub fn set_snowball_capacity_preference(
self,
input: Option<SnowballCapacity>
) -> Self

pub fn shipping_option(self, input: ShippingOption) -> Self

The shipping speed for this job. This speed doesn't dictate how soon you'll get the Snow device, rather it represents how quickly the Snow device moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

pub fn set_shipping_option(self, input: Option<ShippingOption>) -> Self

pub fn notification(self, input: Notification) -> Self

Defines the Amazon Simple Notification Service (Amazon SNS) notification settings for this job.

pub fn set_notification(self, input: Option<Notification>) -> Self

pub fn cluster_id(self, input: impl Into<String>) -> Self

The ID of a cluster. If you're creating a job for a node in a cluster, you need to provide only this clusterId value. The other job attributes are inherited from the cluster.

pub fn set_cluster_id(self, input: Option<String>) -> Self

pub fn snowball_type(self, input: SnowballType) -> Self

The type of AWS Snow Family device to use for this job.

For cluster jobs, AWS Snow Family currently supports only the EDGE device type.

The type of AWS Snow device to use for this job. Currently, the only supported device type for cluster jobs is EDGE.

For more information, see Snowball Edge Device Options in the Snowball Edge Developer Guide.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

pub fn set_snowball_type(self, input: Option<SnowballType>) -> Self

pub fn forwarding_address_id(self, input: impl Into<String>) -> Self

The forwarding address ID for a job. This field is not supported in most Regions.

pub fn set_forwarding_address_id(self, input: Option<String>) -> Self

pub fn tax_documents(self, input: TaxDocuments) -> Self

The tax documents required in your AWS Region.

pub fn set_tax_documents(self, input: Option<TaxDocuments>) -> Self

pub fn device_configuration(self, input: DeviceConfiguration) -> Self

Defines the device configuration for an AWS Snowcone job.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

pub fn set_device_configuration(
self,
input: Option<DeviceConfiguration>
) -> Self

pub fn remote_management(self, input: RemoteManagement) -> Self

Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

pub fn set_remote_management(self, input: Option<RemoteManagement>) -> Self

pub fn long_term_pricing_id(self, input: impl Into<String>) -> Self

The ID of the long-term pricing type for the device.

pub fn set_long_term_pricing_id(self, input: Option<String>) -> Self

pub fn build(self) -> Result<CreateJobInput, BuildError>

Consumes the builder and constructs a CreateJobInput