Module aws_sdk_snowball::model

Modules

address

See Address

cluster_list_entry

See ClusterListEntry

cluster_metadata

See ClusterMetadata

compatible_image

See CompatibleImage

data_transfer

See DataTransfer

device_configuration

See DeviceConfiguration

ec2_ami_resource

See Ec2AmiResource

event_trigger_definition

See EventTriggerDefinition

ind_tax_documents

See IndTaxDocuments

job_list_entry

See JobListEntry

job_logs

See JobLogs

job_metadata

See JobMetadata

job_resource

See JobResource

key_range

See KeyRange

lambda_resource

See LambdaResource

long_term_pricing_list_entry

See LongTermPricingListEntry

nfs_on_device_service_configuration

See NfsOnDeviceServiceConfiguration

notification

See Notification

on_device_service_configuration

See OnDeviceServiceConfiguration

s3_resource

See S3Resource

shipment

See Shipment

shipping_details

See ShippingDetails

snowcone_device_configuration

See SnowconeDeviceConfiguration

target_on_device_service

See TargetOnDeviceService

tax_documents

See TaxDocuments

wireless_connection

See WirelessConnection

Structs

Address

The address that you want the Snow device(s) associated with a specific job to

ClusterListEntry

Contains a cluster's state, a cluster's ID, and other important information.

ClusterMetadata

Contains metadata about a specific cluster.

CompatibleImage

A JSON-formatted object that describes a compatible Amazon Machine Image (AMI),

DataTransfer

Defines the real-time status of a Snow device's data transfer while the device is at AWS.

DeviceConfiguration

The container for SnowconeDeviceConfiguration.

Ec2AmiResource

A JSON-formatted object that contains the IDs for an Amazon Machine Image (AMI),

EventTriggerDefinition

The container for the EventTriggerDefinition$EventResourceARN.

IndTaxDocuments

The tax documents required in AWS Regions in India.

JobListEntry

Each JobListEntry object contains a job's state, a job's ID, and a value

JobLogs

Contains job logs. Whenever a Snow device is used to import data into or export data out of

JobMetadata

Contains information about a specific job including shipping information, job status,

JobResource

Contains an array of AWS resource objects. Each object represents an Amazon S3 bucket,

KeyRange

Contains a key range. For export jobs, a S3Resource object can have an

LambdaResource

Identifies

LongTermPricingListEntry

Each LongTermPricingListEntry object contains information about a long-term pricing type.

NfsOnDeviceServiceConfiguration

An object that represents metadata and configuration settings for NFS service on an AWS Snow Family device.

Notification

The Amazon Simple Notification Service (Amazon SNS) notification settings associated

OnDeviceServiceConfiguration

An object that represents metadata and configuration settings for services on an AWS Snow Family device.

S3Resource

Each S3Resource object represents an Amazon S3 bucket that your

Shipment

The Status and TrackingNumber information for an inbound or

ShippingDetails

A job's shipping information, including inbound and outbound tracking numbers and

SnowconeDeviceConfiguration

Specifies the device configuration for an AWS Snowcone job.

TargetOnDeviceService

An object that represents the service or services on the Snow Family device that your

TaxDocuments

The tax documents required in your AWS Region.

WirelessConnection

Configures the wireless connection on an AWS Snowcone device.

Enums

ClusterState

DeviceServiceName

JobState

JobType

LongTermPricingType

RemoteManagement

ShipmentState

ShippingLabelStatus

ShippingOption

SnowballCapacity

SnowballType

StorageUnit

TransferOption