Struct aws_sdk_snowball::model::job_metadata::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for JobMetadata

Implementations

impl Builder

pub fn job_id(self, input: impl Into<String>) -> Self

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

pub fn set_job_id(self, input: Option<String>) -> Self

pub fn job_state(self, input: JobState) -> Self

The current status of the jobs.

pub fn set_job_state(self, input: Option<JobState>) -> Self

pub fn job_type(self, input: JobType) -> Self

The type of job.

pub fn set_job_type(self, input: Option<JobType>) -> Self

pub fn snowball_type(self, input: SnowballType) -> Self

The type of device used with this job.

pub fn set_snowball_type(self, input: Option<SnowballType>) -> Self

pub fn creation_date(self, input: Instant) -> Self

The creation date for this job.

pub fn set_creation_date(self, input: Option<Instant>) -> Self

pub fn resources(self, input: JobResource) -> Self

An array of S3Resource objects. Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into.

pub fn set_resources(self, input: Option<JobResource>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The description of the job, provided at job creation.

pub fn set_description(self, input: Option<String>) -> Self

pub fn kms_key_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the AWS Key Management Service (AWS KMS) key associated with this job. This ARN was created using the CreateKey API action in AWS KMS.

pub fn set_kms_key_arn(self, input: Option<String>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The role ARN associated with this job. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn address_id(self, input: impl Into<String>) -> Self

The ID for the address that you want the Snow device shipped to.

pub fn set_address_id(self, input: Option<String>) -> Self

pub fn shipping_details(self, input: ShippingDetails) -> Self

A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

pub fn set_shipping_details(self, input: Option<ShippingDetails>) -> Self

pub fn snowball_capacity_preference(self, input: SnowballCapacity) -> Self

The Snow device capacity preference for this job, specified at job creation. In US regions, you can choose between 50 TB and 80 TB Snowballs. All other regions use 80 TB capacity Snowballs.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

pub fn set_snowball_capacity_preference(
self,
input: Option<SnowballCapacity>
) -> Self

pub fn notification(self, input: Notification) -> Self

The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

pub fn set_notification(self, input: Option<Notification>) -> Self

pub fn data_transfer_progress(self, input: DataTransfer) -> Self

A value that defines the real-time status of a Snow device's data transfer while the device is at AWS. This data is only available while a job has a JobState value of InProgress, for both import and export jobs.

pub fn set_data_transfer_progress(self, input: Option<DataTransfer>) -> Self

pub fn job_log_info(self, input: JobLogs) -> Self

Links to Amazon S3 presigned URLs for the job report and logs. For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snow device for your job part is being delivered to you.

pub fn set_job_log_info(self, input: Option<JobLogs>) -> Self

pub fn cluster_id(self, input: impl Into<String>) -> Self

The 39-character ID for the cluster, for example CID123e4567-e89b-12d3-a456-426655440000.

pub fn set_cluster_id(self, input: Option<String>) -> Self

pub fn forwarding_address_id(self, input: impl Into<String>) -> Self

The ID of the address that you want a job shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

pub fn set_forwarding_address_id(self, input: Option<String>) -> Self

pub fn tax_documents(self, input: TaxDocuments) -> Self

The metadata associated with the tax documents required in your AWS Region.

pub fn set_tax_documents(self, input: Option<TaxDocuments>) -> Self

pub fn device_configuration(self, input: DeviceConfiguration) -> Self

The container for SnowconeDeviceConfiguration.

pub fn set_device_configuration(
self,
input: Option<DeviceConfiguration>
) -> Self

pub fn remote_management(self, input: RemoteManagement) -> Self

Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

pub fn set_remote_management(self, input: Option<RemoteManagement>) -> Self

pub fn long_term_pricing_id(self, input: impl Into<String>) -> Self

The ID of the long-term pricing type for the device.

pub fn set_long_term_pricing_id(self, input: Option<String>) -> Self

pub fn on_device_service_configuration(
self,
input: OnDeviceServiceConfiguration
) -> Self

Represents metadata and configuration settings for services on an AWS Snow Family device.

pub fn set_on_device_service_configuration(
self,
input: Option<OnDeviceServiceConfiguration>
) -> Self

pub fn build(self) -> JobMetadata

Consumes the builder and constructs a JobMetadata