Struct aws_sdk_snowball::model::JobMetadata

#[non_exhaustive]
pub struct JobMetadata {

    pub job_id: Option<String>,
    pub job_state: Option<JobState>,
    pub job_type: Option<JobType>,
    pub snowball_type: Option<SnowballType>,
    pub creation_date: Option<Instant>,
    pub resources: Option<JobResource>,
    pub description: Option<String>,
    pub kms_key_arn: Option<String>,
    pub role_arn: Option<String>,
    pub address_id: Option<String>,
    pub shipping_details: Option<ShippingDetails>,
    pub snowball_capacity_preference: Option<SnowballCapacity>,
    pub notification: Option<Notification>,
    pub data_transfer_progress: Option<DataTransfer>,
    pub job_log_info: Option<JobLogs>,
    pub cluster_id: Option<String>,
    pub forwarding_address_id: Option<String>,
    pub tax_documents: Option<TaxDocuments>,
    pub device_configuration: Option<DeviceConfiguration>,
    pub remote_management: Option<RemoteManagement>,
    pub long_term_pricing_id: Option<String>,
    pub on_device_service_configuration: Option<OnDeviceServiceConfiguration>,

}

Contains information about a specific job including shipping information, job status, and other important metadata. This information is returned as a part of the response syntax of the DescribeJob action.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

job_id: Option<String>

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

job_state: Option<JobState>

The current status of the jobs.

job_type: Option<JobType>

The type of job.

snowball_type: Option<SnowballType>

The type of device used with this job.

creation_date: Option<Instant>

The creation date for this job.

resources: Option<JobResource>

An array of S3Resource objects. Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into.

description: Option<String>

The description of the job, provided at job creation.

kms_key_arn: Option<String>

The Amazon Resource Name (ARN) for the AWS Key Management Service (AWS KMS) key associated with this job. This ARN was created using the CreateKey API action in AWS KMS.

role_arn: Option<String>

The role ARN associated with this job. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

address_id: Option<String>

The ID for the address that you want the Snow device shipped to.

shipping_details: Option<ShippingDetails>

A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

snowball_capacity_preference: Option<SnowballCapacity>

The Snow device capacity preference for this job, specified at job creation. In US regions, you can choose between 50 TB and 80 TB Snowballs. All other regions use 80 TB capacity Snowballs.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

notification: Option<Notification>

The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

data_transfer_progress: Option<DataTransfer>

A value that defines the real-time status of a Snow device's data transfer while the device is at AWS. This data is only available while a job has a JobState value of InProgress, for both import and export jobs.

job_log_info: Option<JobLogs>

Links to Amazon S3 presigned URLs for the job report and logs. For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snow device for your job part is being delivered to you.

cluster_id: Option<String>

The 39-character ID for the cluster, for example CID123e4567-e89b-12d3-a456-426655440000.

forwarding_address_id: Option<String>

The ID of the address that you want a job shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

tax_documents: Option<TaxDocuments>

The metadata associated with the tax documents required in your AWS Region.

device_configuration: Option<DeviceConfiguration>

The container for SnowconeDeviceConfiguration.

remote_management: Option<RemoteManagement>

Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

long_term_pricing_id: Option<String>

The ID of the long-term pricing type for the device.

on_device_service_configuration: Option<OnDeviceServiceConfiguration>

Represents metadata and configuration settings for services on an AWS Snow Family device.

Implementations

impl JobMetadata

pub fn builder() -> Builder

Creates a new builder-style object to manufacture JobMetadata