Module aws_sdk_snowball::operation

Structs

CancelCluster

Cancels a cluster job. You can only cancel a cluster job while it's in the

CancelJob

Cancels the specified job. You can only cancel a job before its JobState

CreateAddress

Creates an address for a Snow device to be shipped to. In most regions,

CreateCluster

Creates an empty cluster. Each cluster supports five nodes. You use the CreateJob action separately to create the jobs for each of these nodes. The

CreateJob

Creates a job to import or export data between Amazon S3 and your on-premises data

CreateLongTermPricing

Creates a job with the long-term usage option for a device. The long-term usage is a

CreateReturnShippingLabel

Creates a shipping label that will be used to return the Snow device to AWS.

DescribeAddress

Takes an AddressId and returns specific details about that address in the

DescribeAddresses

Returns a specified number of ADDRESS objects. Calling this API in one of

DescribeCluster

Returns information about a specific cluster including shipping information, cluster

DescribeJob

Returns information about a specific job including shipping information, job status,

DescribeReturnShippingLabel

Information on the shipping label of a Snow device that is being returned to AWS.

GetJobManifest

Returns a link to an Amazon S3 presigned URL for the manifest file associated with the

GetJobUnlockCode

Returns the UnlockCode code value for the specified job. A particular

GetSnowballUsage

Returns information about the Snow Family service limit for your account, and also the

GetSoftwareUpdates

Returns an Amazon S3 presigned URL for an update file associated with a specified

ListClusterJobs

Returns an array of JobListEntry objects of the specified length. Each

ListClusters

Returns an array of ClusterListEntry objects of the specified length. Each

ListCompatibleImages

This action returns a list of the different Amazon EC2 Amazon Machine Images (AMIs)

ListJobs

Returns an array of JobListEntry objects of the specified length. Each

ListLongTermPricing

Lists all long-term pricing types.

UpdateCluster

While a cluster's ClusterState value is in the AwaitingQuorum

UpdateJob

While a job's JobState value is New, you can update some of

UpdateJobShipmentState

Updates the state when a shipment state changes to a different state.

UpdateLongTermPricing

Updates the long-term pricing type.